+++ CHANGELOG.md
@@ -8,14 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 https://github.com/nwnxee/unified/compare/build8193.13...HEAD
 
 ### Added
-N/A
+- ServerLogRedirector: added environment variable `NWNX_SERVERLOGREDIRECTOR_HIDE_VALIDATEGFFRESOURCE_MESSAGES` to hide `*** ValidateGFFResource sent by user.` messages from the NWNX log.
 
 ##### New Plugins
 N/A
 
 ##### New NWScript Functions
-- Object: GetDoorHasVisibleModel()
 - Creature: {Get|Set}CriticalMultiplier{Modifier|Override}() and {Get|Set}CriticalRange{Modifier|Override}()
+- Object: GetDoorHasVisibleModel()
+- Object: GetIsDestroyable()
+- Player: SetCustomToken()
 
 ### Changed
 N/A